requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.1.0
blinker==1.7.0
boto3==1.35.1
botocore==1.35.1
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
clickhouse-connect==0.7.19
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.2
gitdb==4.0.11
idna==3.7
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
Jinja2==3.1.3
jmespath==1.0.1
keyring==24.3.1
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
minio==7.2.8
more-itertools==10.4.0
msgpack==1.0.8
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
protobuf==5.27.3
ptyprocess==0.7.0
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydeck==0.9.1
Pygments==2.18.0
pyproject_hooks==1.1.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.3
rapidfuzz==3.9.6
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.1
s3transfer==0.10.2
schedule==1.2.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
tornado==6.4.1
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
watchdog==4.0.1
Werkzeug==3.0.2
zstandard==0.23.0